+++ lib/config/routes_manager.dart
@@ -8,8 +8,6 @@ import 'package:kafa2a/features/home/user/presentation/screens/user_main_screen.
 import 'package:kafa2a/features/offers/user/presentation/screens/user_offers_screen.dart';
 import 'package:kafa2a/features/onboarding/onboarding_screen.dart';
 import 'package:kafa2a/features/suspension/provider_suspension_main_screen.dart';
-import 'package:kafa2a/features/requests/user/presentation/screens/request_details_screen.dart';
-import 'package:kafa2a/features/requests/user/presentation/screens/all_offers_screen.dart';
 
 class RoutesManager {
   static const String onBoarding = '/onboarding';
@@ -64,16 +62,6 @@ class RoutesManager {
       case homeSuspendedProvider:
         return MaterialPageRoute(
           builder: (context) => ProviderSuspensionMainScreen(),
-        );
-      case requestDetails:
-        final int requestId = settings.arguments as int;
-        return MaterialPageRoute(
-          builder: (context) => RequestDetailsScreen(requestId: requestId),
-        );
-      case allOffers:
-        final int requestId = settings.arguments as int;
-        return MaterialPageRoute(
-          builder: (context) => AllOffersScreen(requestId: requestId),
         );
     }
     return null;
