+++ lib/features/suspension/provider_suspension_main_screen.dart
@@ -1,9 +1,7 @@
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter/material.dart';
 import 'package:kafa2a/core/widgets/custom_navigation_bar.dart';
 import 'package:kafa2a/features/my_profile/presentation/screens/provider/provider_profile_screen.dart';
 import 'package:kafa2a/features/suspension/provider_suspension_home_tab.dart';
-import 'package:kafa2a/l10n/languages/app_localizations.dart';
 
 class ProviderSuspensionMainScreen extends StatefulWidget {
   const ProviderSuspensionMainScreen({super.key});
