+++ lib/l10n/languages/app_localizations_ar.dart
@@ -331,4 +331,7 @@ class AppLocalizationsAr extends AppLocalizations {
 
   @override
   String get paid => 'تم الدفع';
+
+  @override
+  String get allRequests => 'كل الطلبات';
 }
