+++ lib/l10n/languages/app_localizations_en.dart
@@ -331,4 +331,7 @@ class AppLocalizationsEn extends AppLocalizations {
 
   @override
   String get paid => 'Paid';
+
+  @override
+  String get allRequests => 'All Requests';
 }
